+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Update select and download feature flag key to v2
 - Bump Sentry to v6.3.0
 - Bump Sentry gradle plugin to v3.1.3
+- [In Progress: 26 Jul 2022] Load search suggestions when the search query is changed
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-07-25-8342